Use Int instead of Byte

+++ examples/arduino-uno/multiple-blinks.ts
@@ -1,7 +1,7 @@
-import { Byte, periodic } from '@amnisio/rivulet';
+import { Int, periodic } from '@amnisio/rivulet';
 import { Sources, HIGH, LOW, run, createSinks } from '@amnisio/arduino-uno';
 
-const invert = (value: Byte) => value == LOW ? HIGH : LOW;
+const invert = (value: Int) => value == LOW ? HIGH : LOW;
 
 // Sample application that will blink multiple LEDs attached to the arduino UNO at various cycles.
 // Requires an LED to be connected at pin D2.
